backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
pytest
email-validator
python-jose
passlib[bcrypt]
sqlalchemy
